Add VM tests to bbuildbot builder.
Parameterize lunch target and emulator binary names in the build config.

Add a RunLunchWrapper that setups up the lunch ENVs before invoking a
command.

Add a new stage that starts the brillo emulator as a background process,
waits for "adb devices" to report it's serial number, then runs tests
against the emulator.

BUG=chromium:570810
TEST=bin/cbuildbot --nobootstrap --noreexec --buildbot
     --debug --buildroot ~/tmp/buildroot bbuildbot

Change-Id: I16e355f43275df204b0fdbb2aa0d9d00baf8d679